requirements.txt
colorama==0.4.6
ffmpeg-python==0.2.0
future==0.18.2
pytube==12.1.0